refactor(api): split identities.js controller into per-responsibility modules (#599)
identities.js had grown to 738 lines (audit finding C1). Split it by
responsibility into focused routers under routes/identities/:

  list.js      — /api/identities + /api/identity-columns
  detail.js    — /api/identities/:id (+ contexts/assignments/account-matrix) + /api/identities/by-user/:userId
  overrides.js — PUT/DELETE /api/identities/:id/members/:userId/override
  shared.js    — useSql / db / UUID_RE / hasTable + the exported enrichMembers

identities.js is now a barrel that composes the three routers and
re-exports enrichMembers (imported by identities.enrich.test.js), so
app.js and every public path are unchanged. Pure code move.

Grandfathered complexity entries relocated (no increase): cyclomatic
[30,21] -> list [30] + detail [21]; cognitive [33,25,20] -> list [33,20]
+ detail [25].

No new tests needed — the three existing identities test files already
cover the moved code at 89%+ diff coverage, clearing the 80% gate. Full
API suite green (1752 tests); coverage above every floor; both complexity
gates and the OpenAPI-drift guard pass.
