+++ .ci/cognitive-baseline.json
@@ -88,9 +88,11 @@
     "app/api/src/routes/details/user.js": [
       33
     ],
-    "app/api/src/routes/identities.js": [
+    "app/api/src/routes/identities/detail.js": [
+      25
+    ],
+    "app/api/src/routes/identities/list.js": [
       33,
-      25,
       20
     ],
     "app/api/src/routes/matrix/data.js": [
